refactor: select 사용하여 데이터 가공 후 커스텀 훅으로 분리
