+++ src/pages/main/kanban/KanbanBoard.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from 'react';
 import { useAppSelector } from 'redux/store';
-import { modalModeType } from 'hooks/feature/useModal';
+import { modalModeType } from 'hooks/feature/useApplicationModal';
 import { processTypeList } from 'constants/process';
 import { IApplication, IKabanData } from 'types/interfaces/KanbanProcess';
 import * as S from './KanbanStyledComponents';
